Authorization of trashes and caches

+++ kskp/store/auth/role.py
@@ -1,11 +1,35 @@
 import os
 import uuid
 from sqlalchemy import Column, String, text
-from sqlalchemy.dialects.postgresql import INTEGER, TIMESTAMP, UUID
+from sqlalchemy.dialects.postgresql import INTEGER, BOOLEAN, TIMESTAMP, UUID
+from kskp.core import Constraints
 from kskp.store import BaseModel
 from .user_role import UserRole
 
 class Role(BaseModel):
+
+    class Member():
+        """
+        Userとロールへの参加タイプを纏める
+        """
+        def __init__(self, user, owner=False):
+            self.user = user
+            self.owner = owner 
+
+        def to_json(self):
+            ret = self.user.to_json()
+            ret.update({'owner':self.owner})
+            return ret
+
+        def __repr__(self):
+            return f'Member({self.user.id}, {self.user.name}, {self.owner})'
+
+        def __eq__(self, other):
+            return self.user == other.user and self.owner == other.owner
+
+        def __ne__(self, other):
+            return self.user != other.user or self.owner != other.owner
+
     # テーブル名の定義
     __tablename__ = 'roles'
 
@@ -18,6 +42,7 @@ class Role(BaseModel):
     id           = Column(INTEGER, primary_key=True, autoincrement=True)
     uuid         = Column(UUID, nullable=False, unique=True)
     name         = Column(String, nullable=False)
+    _delete_on_isolated = Column('delete_on_isolated', BOOLEAN, nullable=False)
     _creator_id  = Column('creator', INTEGER)
     _modifier_id = Column('modifier', INTEGER)
     created_at   = Column(TIMESTAMP, default=text('statement_timestamp()'))
@@ -30,7 +55,7 @@ class Role(BaseModel):
     EVERYONE_ROLE_UUID   = 'ee16239b-5ffd-447c-9d05-411906ad7364'
     EVERYONE_ROLE_LABEL  = 'EVERYONE'
 
-    def __init__(self, session, name):
+    def __init__(self, session, name, delete_on_isolated=False):
         """
         コンストラクタ
         """
@@ -40,7 +65,12 @@ class Role(BaseModel):
         # UUIDを採番する
         self.uuid = str(uuid.uuid4())
 
+        # ロール名
         self.name = name
+
+        # Trueの場合、紐づくDatumが存在しなくなったらこのロールを削除する
+        # (このロールの所有者以外のユーザがDatumを削除した時でも削除する)
+        self._delete_on_isolated = delete_on_isolated
 
         # creator, modifier
         if session is not None and session.user is not None:
@@ -73,6 +103,10 @@ class Role(BaseModel):
         return Util.datetime_to_local_time_str(self.created_at)
 
     @property
+    def delete_on_isolated(self):
+        return self._delete_on_isolated
+
+    @property
     def is_sys_admin(self):
         return self.uuid == Role.SYS_ADMIN_ROLE_UUID
 
@@ -84,6 +118,10 @@ class Role(BaseModel):
     def is_everyone(self):
         return self.uuid == Role.EVERYONE_ROLE_UUID
 
+    @property
+    def is_system_role(self):
+        return self.is_sys_admin or self.is_usr_admin or self.is_everyone
+
     def _get_system_role_label(self):
         if self.is_sys_admin:
             return Role.SYS_ADMIN_ROLE_LABEL
@@ -93,6 +131,17 @@ class Role(BaseModel):
             return Role.EVERYONE_ROLE_LABEL
         else:
             return ''
+
+    def raise_no_role_owner_exception(self):
+        from .exceptions import NoRoleOwnerException
+        if self.is_sys_admin:
+            raise NoRoleOwnerException('システム管理者権限を持つユーザがいなくなるのでこの操作はできません')
+        elif self.is_usr_admin:
+            raise NoRoleOwnerException('ユーザ管理者権限を持つユーザがいなくなるのでこの操作はできません')
+        elif self.is_everyone:
+            raise NoRoleOwnerException('ユーザ管理者権限を持つユーザがいなくなるのでこの操作はできません')
+        else:
+            raise NoRoleOwnerException('ロール所有者がいなくなるのでこの操作はできません')
 
     def save(self):
         """
@@ -123,9 +172,9 @@ class Role(BaseModel):
     def delete(self):
         from .auth import Auth
 
-        # ロールに一人以上のユーザが所属している場合は例外を送出する
-        if self.has_joined_user() > 0:
-            raise Exception('Can not delete the role that has user(s).')
+        # # ロールに一人以上のユーザが所属している場合は例外を送出する
+        # if self.has_joined_user() > 0:
+        #     raise Exception('Can not delete the role that has user(s).')
 
         # 
         # 削除によってどのロールからも所有されなくなるデータがある場合は例外を送出する
@@ -135,18 +184,26 @@ class Role(BaseModel):
         # count = self._session.query(Auth).filter(Auth.role_id == self.id)\
         #                                  .filter(Auth.own == 1).count()
 
+        if self.is_system_role:
+            raise Exception(f'ロール({self.name})はシステムロールなので削除できません')
+
         try:
+            self_id = self.id
             # 削除ロールに対する権限情報をauthsテーブルから全て削除する
             self._session.query(Auth).filter(Auth.role_id==self.id).delete()
             # ロールを削除する
             self._session.delete(self)
+            # 削除ロールから全てのユーザを脱退させる
+            # (ロールの削除はロールの所有権が必要なので、
+            #  所有権フラグを持つUserRoleはロール削除の後に削除すること)
+            self._session.query(UserRole).filter(UserRole.role_id==self_id).delete()
         except Exception as e:
             self._session.rollback()
             raise e
         finally:
             self._session.commit()
 
-    def is_self_role(self):
+    def is_self_role(self) -> bool:
         """
         本人ロールの場合はTrueを返す
         """
@@ -154,17 +211,33 @@ class Role(BaseModel):
         count = self._session.query(User).filter(User.self_role_id==self.id).count()
         return count > 0
 
-    def is_joined_user(self, user):
+    def is_joined_user(self, user) -> bool:
         from .user import User
         count1 = self._session.query(User).filter(User.self_role_id==self.id).filter(User.id==user.id).count()
         count2 = self._session.query(UserRole).filter(UserRole.role_id==self.id).filter(UserRole.user_id==user.id).count()
         return count1 + count2 > 0
 
-    def has_joined_user(self):
+    def count_joined_users(self) -> int:
         from .user import User
         count1 = self._session.query(User).filter(User.self_role_id==self.id).count()
         count2 = self._session.query(UserRole).filter(UserRole.role_id==self.id).count()
-        return count1 + count2 > 0
+        return count1 + count2
+
+    def is_owner(self, user) -> bool:
+        query = self._session.query(UserRole).\
+                              filter(UserRole.role_id==self.id).\
+                              filter(UserRole.user_id==user.id).\
+                              filter(UserRole.owner==True)
+        return query.count() > 0
+
+    def count_owners(self) -> int:
+        query = self._session.query(UserRole).\
+                              filter(UserRole.role_id==self.id).\
+                              filter(UserRole.owner==True)
+        return query.count()
+
+    def has_joined_user(self) -> bool:
+        return self.count_joined_users() > 0
 
     def get_joined_users(self):
         """
@@ -175,56 +248,154 @@ class Role(BaseModel):
         from .user import User
 
         exists_user_role = exists().where(and_(UserRole.role_id==self.id, UserRole.user_id==User.id))
-        exists_user = exists().where(and_(User.self_role_id==self.id, User.id==User.id))
 
         query = self._session.query(User).\
-                              filter(or_(exists_user_role, exists_user))
+                              filter(or_(exists_user_role, User.self_role_id==self.id))
                               
         return query.order_by(User.id).all()
 
-    def join_user(self, user):
+    def get_joined_members(self):
+        """
+        ロールに所属する全てのメンバを返す
+        (ユーザID順で返す)
+        """
+        from .user import User
+
+        query = self._session.query(User, UserRole.owner).\
+                              outerjoin(UserRole, UserRole.user_id==User.id).\
+                              filter(UserRole.role_id==self.id)
+        results =  query.order_by(User.id).all()
+
+        members = []
+        for result in results:
+            user = result[0]
+            # Queryクラスは、query(User, ...)の結果にsessionを設定しないのでここで設定する
+            user._session = self._session
+            members.append(Role.Member(user, owner=result[1]))
+
+        return members
+
+    def is_last_owner(self, user):
+        """
+        指定されたユーザがただ一人の所有者ならTrueを返す
+        """
+        return self.is_owner(user) and self.count_owners() <= 1
+
+    def owner_exists(self, members):
+        """
+        指定されたメンバリストに所有者が存在する場合はTrueを返す
+        """
+        for member in members:
+            if member.owner:
+                return True
+        return False
+
+    def join_member(self, member):
         """
         ロールにユーザを所属させる
         """
         if self.is_self_role():
             raise Exception('本人ロールに本人以外のユーザを所属させることはできません')
 
-        if not self.is_joined_user(user):
-            user_role = UserRole(self._session, user.id, self.id)
+        from kskp.store.factory import UserRoleFactory
+        factory = UserRoleFactory(self._session)
+
+        if factory.exists(member.user.id, self.id):
+            # システム系ロールが、この所属によって、ロールに所有者が居なくなる場合はエラーとする
+            if self.is_system_role and member.owner == False and self.is_last_owner(member.user):
+                self.raise_no_role_owner_exception()
+
+            # 既にメンバの場合は所有権フラグを更新する
+            user_role = factory.find_by_id(member.user.id, self.id)
+            user_role.update_owner(member.owner)
+        else:
+            # ロールにメンバを追加する
+            user_role = UserRole(self._session, member.user.id, self.id, member.owner)
             user_role.save()
-    
-    def leave_user(self, user):
+
+    def leave_member(self, user):
         """
         ロールからユーザを脱退させる
         """
         if self.is_self_role():
             raise Exception('本人ロールからユーザを脱退させることはできません')
 
-        if self.is_joined_user(user):
-            from kskp.store.factory import UserRoleFactory
-            user_role = UserRoleFactory(self._session).find_by_id(user.id, self.id)
+        from kskp.store.factory import UserRoleFactory
+        factory = UserRoleFactory(self._session)
+
+        if factory.exists(user.id, self.id):
+            # システム系ロールが、この脱退によって、ロールに所有者が居なくなる場合はエラーとする
+            if self.is_system_role and self.is_last_owner(user):
+                self.raise_no_role_owner_exception()
+            # ロールからメンバを削除する
+            user_role = factory.find_by_id(user.id, self.id)
             user_role.delete()
 
-    def leave_all_users(self):
+    def leave_others(self, except_user):
         """
-        ロールから全てのユーザを脱退させる
+        指定する1人を除いて、ロールから他のユーザを全て脱退させる
         """
         if self.is_self_role():
             raise Exception('本人ロールからユーザを脱退させることはできません')
 
-        from kskp.store.factory import UserRoleFactory
-        UserRoleFactory(self._session).delete_all_by_role_id(self.id)
+        # システム系ロールが、この脱退によって、ロールに所有者が居なくなる場合はエラーとする
+        if self.is_system_role and not self.is_owner(except_user):
+            self.raise_no_role_owner_exception()
 
-    def init_users(self, users):
+        from kskp.store.factory import UserRoleFactory
+        UserRoleFactory(self._session).delete_all_by_role_id(self.id, except_user_id=except_user.id)
+
+    def init_members(self, members):
         """
         ロールの所属ユーザを初期化する
         """
-        # 一旦、全てのユーザを削除する
-        self.leave_all_users()
+        from .exceptions import NotAuthorizedException
 
-        # ユーザを追加する
-        for user in users:
-            self.join_user(user)
+        # 
+        # 指定されたメンバリストの妥当性を検証する
+        # 
+        users = set()
+        owner_exists = False
+        for member in members:
+            # ロール所有者が設定されない場合はエラーとする
+            if member.owner:
+                owner_exists = True
+
+            # 1人のUserが重複指定された場合はエラーとする
+            if member.user in users:
+                raise Exception(f'ユーザ({member.user.name})が重複して指定されました')
+            else:
+                users.add(member.user)
+
+        # システム系ロールが、この初期化によって、ロールに所有者が居なくなる場合はエラーとする
+        if self.is_system_role and not owner_exists:
+            self.raise_no_role_owner_exception()
+
+        # 操作ユーザがロール所有者以外の場合はエラーとする
+        self_user = self._session.user
+        if not self.is_owner(self_user):
+            raise NotAuthorizedException(f'ロール所有者以外のユーザ({self_user})は所属ユーザの初期化をできません')
+
+        # ロールから、自分以外のユーザを全て削除する
+        for user in self.get_joined_users():
+            if user == self_user:
+                continue
+            self.leave_member(user)
+
+        # 自分以外のユーザを全て設定する
+        self_member = None
+        for member in members:
+            if member.user == self_user:
+                self_member = member
+            else:
+                self.join_member(member)
+
+        if self_member is None:
+            # 自分がメンバに指定されていない場合は自分を削除する
+            self.leave_member(self_user)
+        else:
+            # 自分がメンバに指定されている場合は改めて追加する
+            self.join_member(self_member)
 
     def init_authz(self, datum_id, read, write, exec=None, own=None):
         from .auth import Auth
@@ -256,13 +427,8 @@ class Role(BaseModel):
             # permission=Noneが指定された場合は何もしない
             pass
         else:
-            authz = auth_factory.create(self.id, datum_id, operation, permission=permission)
-            authz.save()
-
-    @property
-    def created_at_str(self):
-        from kskp.core import Util
-        return Util.datetime_to_local_time_str(self.created_at)
+            auth = auth_factory.create(self.id, datum_id, operation, permission=permission)
+            auth.save()
 
     def to_json(self):
         return {
@@ -281,3 +447,6 @@ class Role(BaseModel):
 
     def __ne__(self, other):
         return self.uuid != other.uuid
+
+    def __hash__(self) -> int:
+        return hash(self.uuid)
